Add user activity history dummy table
