+++ setup.cfg
@@ -37,7 +37,7 @@ edx.analytics.tasks =
     database-import = edx.analytics.tasks.database_imports:ImportAllDatabaseTablesTask
     location-per-course = edx.analytics.tasks.location_per_course:LastCountryOfUser
     course-facts-to-mysql = edx.analytics.tasks.daily_enrollment_trends:ImportCourseDailyFactsIntoMysql
-    course-fact = edx.analytics.tasks.course:CourseFactTask
+    course-fact = edx.analytics.tasks.course_fact:CourseFactTask
 
 mapreduce.engine =
     hadoop = edx.analytics.tasks.mapreduce:MapReduceJobRunner
